Actualization-V0.0.9

+++ Assets/Scripts/GameManager.cs
@@ -9,7 +9,7 @@ public class GameManager : MonoBehaviour
     public float gameSpeed = 1;
     public bool trapSpawned = false;
     public float timeBeforNextSpawn = 3000f;
-    public int limpieza = 0;
+    public float limpieza = 0f;
 
     void Awake()
     {
